feat: added support for multiple ports

+++ run_container.sh
@@ -1,13 +1,24 @@
 #!/usr/bin/env bash
 
-set -e 
+set -e
 
-PORT=3490
+PORTS=("81:81" "85:85" "3490:3490" "3491:3491")
 CONTAINER_NAME="ciaomamma"
 
-if [[ $(docker ps | grep tester) ]]; then
+if [[ $(docker ps -a --filter "name=^/${CONTAINER_NAME}$" --format '{{.Names}}') == $CONTAINER_NAME ]]; then
+    docker rm $CONTAINER_NAME
+fi
+
+if [[ $(docker ps -a | grep tester) ]]; then
     docker kill $CONTAINER_NAME
 fi
 
 docker build -t webserv .
-docker run --name $CONTAINER_NAME -dp $PORT:$PORT webserv
+
+# Build the ports arguments string
+PORTS_ARGS=""
+for port in "${PORTS[@]}"; do
+  PORTS_ARGS+=" -p ${port}"
+done
+
+docker run --name $CONTAINER_NAME -d ${PORTS_ARGS} webserv
